Reworked position mappers (idk, there were same mappers in different files)
