+++ src/CompanyService.Mappers/RequestMappers/Interfaces/IPositionMapper.cs
@@ -1,0 +1,19 @@
+﻿using LT.DigitalOffice.CompanyService.Mappers.Interfaces;
+using LT.DigitalOffice.CompanyService.Models.Db;
+using LT.DigitalOffice.CompanyService.Models.Dto.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace LT.DigitalOffice.CompanyService.Mappers.RequestMappers.Interfaces
+{
+    /// <summary>
+    /// Represents mapper. Provides methods for converting an object of <see cref="Position"/>
+    /// type into an object of <see cref="DbPosition"/> type according to some rule.
+    /// </summary>
+    public interface IPositionMapper : IMapper<Position, DbPosition>
+    {
+    }
+}
