+++ src/CompanyService.Mappers/RequestMappers/PositionMapper.cs
@@ -1,0 +1,30 @@
+﻿using LT.DigitalOffice.CompanyService.Mappers.RequestMappers.Interfaces;
+using LT.DigitalOffice.CompanyService.Models.Db;
+using LT.DigitalOffice.CompanyService.Models.Dto.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace LT.DigitalOffice.CompanyService.Mappers.RequestMappers
+{
+    public class PositionMapper : IPositionMapper
+    {
+        public DbPosition Map(Position value)
+        {
+            if (value == null)
+            {
+                throw new ArgumentNullException(nameof(value));
+            }
+
+            return new DbPosition
+            {
+                Id = Guid.NewGuid(),
+                Name = value.Name,
+                Description = value.Description,
+                IsActive = value.IsActive
+            };
+        }
+    }
+}
