+++ test/CompanyService.Mappers.UnitTests/PositionMapperTests.cs
@@ -1,4 +1,6 @@
 ﻿using LT.DigitalOffice.CompanyService.Mappers.Interfaces;
+using LT.DigitalOffice.CompanyService.Mappers.RequestMappers;
+using LT.DigitalOffice.CompanyService.Mappers.ResponsesMappers;
 using LT.DigitalOffice.CompanyService.Models.Db;
 using LT.DigitalOffice.CompanyService.Models.Dto;
 using LT.DigitalOffice.CompanyService.Models.Dto.Models;
@@ -26,7 +28,7 @@ namespace LT.DigitalOffice.CompanyService.Mappers.UnitTests
         [OneTimeSetUp]
         public void OneTimeSetUp()
         {
-            mapperDbPositionToPosition = new PositionMapper();
+            mapperDbPositionToPosition = new DbPositionMapper();
             mapperEditPositionRequestToDbPosition = new PositionMapper();
             mapperAddPositionRequest = new PositionMapper();
         }
